cp: split IDefaultComponent from IComponent to clarify the development of std component's default impl
